requested fixes including copyright headers and analytic class expansion
