+++ analytics/analytics-ui/src/app/analytics/analytic-filter.pipe.ts
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2019 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { Pipe, PipeTransform } from '@angular/core';
+
+class Analytic {
+    analyticName: string;
+    operationName: string;
+    description: string;
+    userId: string;
+    readers: Array<string>;
+    writers: Array<string>;
+    uiMapping: Map<string, string>;
+    metaData: string;
+    outputVisualisation: string;
+    score: number;
+    options: Map<string, string>;
+}
+
+@Pipe({
+    name: 'analyticFilter'
+})
+export class AnalyticFilterPipe implements PipeTransform {
+    transform(items: Analytic[], searchText: string): Analytic[] {
+        if (!items) { return []; }
+        if (!searchText) { return items; }
+        searchText = searchText.toLowerCase();
+        return items.filter(it => {
+            return it.analyticName.toLowerCase().includes(searchText);
+        });
+    }
+}
