+++ analytics/analytics-ui/src/app/analytics/analytic/analytic.component.css
@@ -18,6 +18,7 @@ mat-card {
   display: flex;
   align-items: flex-end;
   justify-content: center;
+  height: 100%
 }
 
 .scalableFont {
