+++ analytics/analytics-ui/src/app/analytics/analytic/analytic.component.html
@@ -1,6 +1,6 @@
 <mat-card id="{{model.operationName}}" (click)="execute(model)" class="grid-button" md-ink-ripple
   matTooltip="{{model.description}}" style.background="{{model.metaData.colour}}">
-  <div style="height: 100%" fxLayout="column" fxLayoutAlign="center">
+  <div fxLayout="column" fxLayoutAlign="center">
     <div fxFlex="30%" class="titleContainer">
       <h2 class="scalableFont">{{model.analyticName}}</h2>
     </div>
