+++ analytics/analytics-ui/src/app/parameters/parameter-page/parameter-page.component.css
@@ -1,0 +1,7 @@
+.title {
+  margin-left: 110px
+}
+
+.execute {
+  margin-right: 110px
+}
