+++ analytics/analytics-ui/src/app/parameters/parameter-page/parameter-page.component.html
@@ -1,4 +1,4 @@
-<div style="margin-left:110px">
+<div class="title">
   <h1>Parameters for {{analytic.analyticName}}</h1>
 </div>
 <form name="parameterPageForm" #parameterPageForm="ngForm">
@@ -6,7 +6,7 @@
     <app-query [model]="analytic">
     </app-query>
   </mat-card>
-  <div style="margin-right:110px" fxLayout="row" fxLayoutAlign="end none">
+  <div fxLayout="row" fxLayoutAlign="end none" class="execute">
     <mat-progress-spinner *ngIf="loading" class="example-margin" style="margin-right:10px" [color]="color" [mode]="mode"
       [diameter]="40">
     </mat-progress-spinner>
